Changes as per review comments
