+++ src/dotnet/APIView/APIView/Languages/CodeFileBuilder.cs
@@ -11,14 +11,12 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.IO;
 using System.Linq;
-using System.Text.RegularExpressions;
 
 namespace ApiView
 {
     public class CodeFileBuilder
     {
         private static readonly char[] _newlineChars = new char[] { '\r', '\n' };
-        private static Regex _packageNameParser = new Regex("([A-Za-z.]*[a-z]).([\\S]*)", RegexOptions.Compiled);
 
         SymbolDisplayFormat _defaultDisplayFormat = new SymbolDisplayFormat(
             SymbolDisplayGlobalNamespaceStyle.Omitted,
@@ -70,71 +68,48 @@ namespace ApiView
             }
         }
 
-        public CodeFile Build(IAssemblySymbol assemblySymbol, bool runAnalysis, string originalName)
+        public CodeFile Build(IAssemblySymbol assemblySymbol, bool runAnalysis)
         {
+            _assembly = assemblySymbol;
+            var analyzer = new Analyzer();
+
+            if (runAnalysis)
+            {
+                analyzer.VisitAssembly(assemblySymbol);
+            }
             var builder = new CodeFileTokensBuilder();
             var navigationItems = new List<NavigationItem>();
-            var analyzer = new Analyzer();
-            var packageName = "";
-            var reviewName = "";
-
-            if (assemblySymbol != null)
+            foreach (var namespaceSymbol in SymbolOrderProvider.OrderNamespaces(EnumerateNamespaces(assemblySymbol)))
             {
-                packageName = assemblySymbol.Name;
-                reviewName = $"{assemblySymbol.Name} ({assemblySymbol.Identity.Version})";
-                _assembly = assemblySymbol;
-
-                if (runAnalysis)
+                if (namespaceSymbol.IsGlobalNamespace)
                 {
-                    analyzer.VisitAssembly(assemblySymbol);
-                }
-
-                foreach (var namespaceSymbol in SymbolOrderProvider.OrderNamespaces(EnumerateNamespaces(assemblySymbol)))
-                {
-                    if (namespaceSymbol.IsGlobalNamespace)
+                    foreach (var namedTypeSymbol in SymbolOrderProvider.OrderTypes(namespaceSymbol.GetTypeMembers()))
                     {
-                        foreach (var namedTypeSymbol in SymbolOrderProvider.OrderTypes(namespaceSymbol.GetTypeMembers()))
-                        {
-                            BuildType(builder, namedTypeSymbol, navigationItems);
-                        }
+                        BuildType(builder, namedTypeSymbol, navigationItems);
                     }
-                    else
-                    {
-                        Build(builder, namespaceSymbol, navigationItems);
-                    }
-                }
-            }
-            else if (originalName != null)
-            {
-                packageName = Path.GetFileNameWithoutExtension(originalName);
-                var packageNameMatch = _packageNameParser.Match(packageName);
-                if (packageNameMatch.Success)
-                {
-                    packageName = packageNameMatch.Groups[1].Value;
-                    reviewName = $"{packageName} ({packageNameMatch.Groups[2].Value})";
                 }
                 else
                 {
-                    reviewName = $"{packageName} (metapackage)";
+                    Build(builder, namespaceSymbol, navigationItems);
                 }
             }
 
             NavigationItem assemblyNavigationItem = new NavigationItem()
             {
-                Text = packageName + ".dll",
+                Text = assemblySymbol.Name + ".dll",
                 ChildItems = navigationItems.ToArray(),
                 Tags = { { "TypeKind", "assembly" } }
             };
 
             var node = new CodeFile()
             {
-                Name = reviewName,
+                Name = $"{assemblySymbol.Name} ({assemblySymbol.Identity.Version})",
                 Language = "C#",
                 Tokens = builder.Tokens.ToArray(),
                 VersionString = CurrentVersion,
                 Navigation = new[] { assemblyNavigationItem },
                 Diagnostics = analyzer.Results.ToArray(),
-                PackageName = packageName
+                PackageName = assemblySymbol.Name
             };
 
             return node;
