+++ src/dotnet/APIView/APIViewWeb/Languages/CSharpLanguageService.cs
@@ -4,6 +4,7 @@
 using System;
 using System.IO;
 using System.IO.Compression;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using ApiView;
 
@@ -11,6 +12,7 @@ namespace APIViewWeb
 {
     public class CSharpLanguageService : LanguageService
     {
+        private static Regex _packageNameParser = new Regex("([A-Za-z.]*[a-z]).([\\S]*)", RegexOptions.Compiled);
         public override string Name { get; } = "C#";
         public override string Extension { get; } = ".dll";
 
@@ -61,12 +63,41 @@ namespace APIViewWeb
                 }
 
                 var assemblySymbol = CompilationFactory.GetCompilation(dllStream, docStream);
-                return Task.FromResult(new CodeFileBuilder().Build(assemblySymbol, runAnalysis, originalName));
+                if ( assemblySymbol == null)
+                {
+                    return Task.FromResult(GetDummyReviewCodeFile(originalName));
+                }
+                
+                return Task.FromResult(new CodeFileBuilder().Build(assemblySymbol, runAnalysis));
             }
             finally
             {
                 archive?.Dispose();
             }
+        }
+
+        private CodeFile GetDummyReviewCodeFile(string originalName)
+        {
+            var packageName = Path.GetFileNameWithoutExtension(originalName);
+            var reviewName = "";
+            var packageNameMatch = _packageNameParser.Match(packageName);
+            if (packageNameMatch.Success)
+            {
+                packageName = packageNameMatch.Groups[1].Value;
+                reviewName = $"{packageName} ({packageNameMatch.Groups[2].Value})";
+            }
+            else
+            {
+                reviewName = $"{packageName} (metapackage)";
+            }
+
+            return new CodeFile()
+            {                
+                Name = reviewName,
+                Language = "C#",
+                VersionString = CodeFileBuilder.CurrentVersion,
+                PackageName = packageName
+            };
         }
     }
 }
